Extract and move ServerListModel
