+++ common/src/main/java/com/wynntils/features/user/LobbyUptimeFeature.java
@@ -6,8 +6,8 @@ package com.wynntils.features.user;
 
 import com.wynntils.core.features.UserFeature;
 import com.wynntils.core.managers.Model;
-import com.wynntils.core.webapi.ServerListModel;
 import com.wynntils.wynn.item.ItemStackTransformModel;
+import com.wynntils.wynn.model.ServerListModel;
 import java.util.List;
 
 public class LobbyUptimeFeature extends UserFeature {
