+++ common/src/main/java/com/wynntils/wynn/item/ServerItemStack.java
@@ -4,8 +4,8 @@
  */
 package com.wynntils.wynn.item;
 
-import com.wynntils.core.webapi.ServerListModel;
 import com.wynntils.wynn.item.parsers.WynnItemMatchers;
+import com.wynntils.wynn.model.ServerListModel;
 import com.wynntils.wynn.objects.profiles.ServerProfile;
 import java.util.ArrayList;
 import java.util.List;
